flask_multistage_docker-compose/test_requirements.txt
pytest>=3.0.7
pytest-flask